#2559. Rename and regroup augmentation types tests
