+++ LanguageFeatures/Augmentation-libraries/augmentation_libraries_lib.dart
@@ -4,6 +4,7 @@
 
 /// @description Common library for augmentation libraries tests
 /// @author sgrekhov22@gmail.com
+
 // SharedOptions=--enable-experiment=macros
 
 library augmentation_libraries_lib;
@@ -13,3 +14,7 @@ class AL {}
 final class FinalClass {}
 
 base class BaseClass {}
+
+sealed class SealedClass {}
+
+class SealedClassImpl implements SealedClass {}
