--- LanguageFeatures/Augmentation-libraries/augmenting_types_A06_t03_lib.dart
@@ -1,33 +1,0 @@
-// Copyright (c) 2024, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-/// @assertion A class, enum, extension type, or mixin augmentation may specify
-/// extends, implements, on, and with clauses (when generally supported). The
-/// types in these clauses are appended to the original declarations clauses of
-/// the same kind, and if that clause did not exist previously then it is added
-/// with the new types. All regular rules apply after this appending process, so
-/// you cannot have multiple extends on a class, or an on clause on an enum, etc
-///
-/// @description Checks that a class, extension type, mixin and enum augment may
-/// specify `implements` clause
-/// @author sgrekhov22@gmail.com
-
-// SharedOptions=--enable-experiment=macros
-
-augment library 'augmenting_types_A06_t03.dart';
-
-augment class C implements I {
-  String get id => "C";
-}
-
-augment extension type ET(I _) implements I {}
-
-augment mixin M implements I {
-  String get id => "M";
-}
-
-augment enum E implements I {
-  augment e1;
-  String get id => "E";
-}
