+++ LanguageFeatures/Augmentation-libraries/augmenting_types_A07_t02_lib.dart
@@ -9,12 +9,31 @@
 /// with the new types. All regular rules apply after this appending process, so
 /// you cannot have multiple extends on a class, or an on clause on an enum, etc
 ///
-/// @description Checks that it is a compile-time error if a class, mixin or
-/// enum augment specifies a base class in an `implements` clause
+/// @description Checks that a class, extension type, mixin and enum augment may
+/// specify an additional `implements` clause
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=macros
 
-augment library 'augmenting_types_A06_t07.dart';
+augment library 'augmenting_types_A07_t02.dart';
 
-augment mixin M on C {}
+interface class I2 {
+  String get id2 => "I2";
+}
+
+augment class C implements I2 {
+  String get id2 => "I2 from C";
+}
+
+augment extension type ET(I1 _) implements I0 {
+  String get id0 => "I0 from ET";
+}
+
+augment mixin M implements I2 {
+  String get id2 => "I2 from M";
+}
+
+augment enum E implements I2 {
+  augment e1;
+  String get id2 => "I2 from E";
+}
