+++ LanguageFeatures/Augmentation-libraries/augmenting_types_A07_t03.dart
@@ -9,28 +9,34 @@
 /// with the new types. All regular rules apply after this appending process, so
 /// you cannot have multiple extends on a class, or an on clause on an enum, etc
 ///
-/// @description Checks that it is a compile-time error if a class, extension
-/// type, mixin or enum augment specifies a base class in an `implements` clause
+/// @description Checks that it is a compile-time error if a class, mixin or
+/// enum augment specifies `implements` clause but doesn't implement this
+/// interface
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=macros
 
-augment library 'augmenting_types_A06_t06.dart';
-import 'augmentation_libraries_lib.dart';
+import augment 'augmenting_types_A07_t03_lib.dart';
 
-augment class C implements BaseClass {}
-//                         ^^^^^^^^^
+interface class I {
+  String get id => "I";
+}
+
+class C {}
+
+mixin M {}
+
+enum E {
+  e1;
+}
+
+class MA = Object with M;
+//    ^^
 // [analyzer] unspecified
 // [cfe] unspecified
 
-augment mixin M implements BaseClass {}
-//                         ^^^^^^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-
-augment enum E implements BaseClass {
-//                        ^^^^^^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-  augment e1;
+main() {
+  print(C);
+  print(MA);
+  print(E);
 }
