+++ LanguageFeatures/Augmentation-libraries/augmenting_types_A08_t02_lib.dart
@@ -15,7 +15,7 @@
 
 // SharedOptions=--enable-experiment=macros
 
-augment library 'augmenting_types_A06_t08.dart';
+augment library 'augmenting_types_A08_t02.dart';
 
 augment mixin M on C {}
 //            ^
